requirements.txt
certifi==2019.6.16
chardet==3.0.4
click==6.7
Django==2.1.1
djangorestframework==3.10.2
feedparser==5.2.1
Flask==1.0.2
idna==2.8
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
peewee==3.3.4
pytz==2019.1
requests==2.22.0
urllib3==1.25.3
Werkzeug==0.14.1
